feat: Improve account management
Improve account switching process to follow expectations of the
sysmodule. We were previously doing this process wrong and patching
friends and act to workaround it, but now that we know how accounts are
managed, we can fix it to work properly and remove the now unnecessary
patches.

The friends checks that were being patched are about ensureing that
there isn't an account loaded before loading a new one. The ACT patches
workaround the removal of uncommitted accounts. We now create accounts
correctly and commit them, but old accounts will be removed when
restaring. This isn't a big deal though, since the PNID isn't
fundamental on the 3DS, so users can just link it again without much
issue.

When booting with the new patches for the first time using Pretendo, the
console will boot with the Pretendo friends accounnt but with the
Nintendo ACT account. To fix the issue, you will need to switch to
Nintendo and back to Pretendo to make and commit a new ACT account.

Technically the ACT and friends patches aren't needed anymore and they
could be covered with HTTP, but the URL patches have been kept for
better compatibility with Citra (now Azahar). Once Azahar implements HLE
for friends and ACT, their respective patches here should be removed.

Error checking is also improved, and now if any of the steps for
switching accounts fail it will be reported on screen, forcing the user
to restart the system. Alongside that, button controls are now shown on
the bottom screen.