feat: Add PNID migration tool
When pressing B on Nimbus, try to commit the ACT account before it gets
removed upon restarting when upgrading from a previous version. Note
that the user will have to upgrade Nimbus while the console is turned
on, such as with using Universal-Updater or updating the files over FTP.
Also add a few convenience features.