feat: Add Nimbus plugin
This plugin controls different networking-related functionality, such as
RCE patches for games, HokakuCTR and a PID logger for the majority of
PIA games, all in one plugin.

At the moment, there are only patches for ACNL. To allow integration of
these patches into other plugins such as Vapecord (which is licensed
under MIT), the patch code is released under the LGPL-3.0 license, as
noted on the source files.